add scss to go with component
