+++ src/components/addCategory/_addCategory.scss
@@ -1,0 +1,44 @@
+$neutral-200: #EFF6E0;
+$neutral-300: #AEC3B0;
+$primary-700: #124559;
+
+
+.btn-add-category {
+    background-color: $neutral-200;
+    border: none;
+    padding: 0.5rem 1rem;
+    font-size: 1.2rem;
+}
+
+.btn-ok, .btn-cancel {
+    font-size: 1rem;
+    border: none;
+    padding: .4rem 1rem;
+    background-color: $neutral-200;
+    border: 2px solid $neutral-200;
+}
+
+.btn-cancel {
+    background-color: $primary-700;
+    border: 2px solid $neutral-200;
+    color: $neutral-200;
+}
+
+.new-category-input-wrapper {
+    display: flex;
+    flex-direction: column;
+    padding: 1rem;
+    max-width: 300px;
+    gap: 1rem;
+    align-items: center;
+    background-color: $primary-700;
+}
+
+.new-category-input-buttons {
+    display: flex;
+    gap: 1rem;
+}
+
+#new-category-text {
+    font-size: 1.2rem;
+}
